Big refactoring. Restructured folder structure of React components to follow M-V-C structure
