+++ imports/controller/actions/detailedView.js
