+++ imports/controller/reducers/favouritesReducer.js
