+++ imports/view/components/CommonDialog.jsx
