+++ imports/view/components/RecipeCards.jsx
@@ -9,8 +9,8 @@ import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
 import CardMedia from "@material-ui/core/CardMedia";
 import Button from "@material-ui/core/Button";
-import { openDetailedView, closeDetailedView } from '../actions/detailedView.js';
-import { setRecipeDetails } from '../actions/recipe.js';
+import { openDetailedView, closeDetailedView } from '../../controller/actions/detailedView.js';
+import { setRecipeDetails } from '../../controller/actions/recipe.js';
 import RecipeDetails from './RecipeDetails';
 import '../style/RecipeCards.css';
 import Favourites from '../../api/favourites';
