+++ imports/view/components/SearchBar.jsx
@@ -4,7 +4,7 @@ import { compose } from 'redux';
 import { makeStyles, withStyles } from '@material-ui/core/styles';
 import MenuItem from '@material-ui/core/MenuItem';
 import TextField from '@material-ui/core/TextField';
-import { updateInput } from "../actions/input.js";
+import { updateInput } from "../../controller/actions/input.js";
 import PropTypes from 'prop-types';
 import ChipInput from 'material-ui-chip-input';
